// types/foundry/common/data/data.mjs/actorData.d.ts
interface ActorDataConstructorData {
	/**
	 * The _id which uniquely identifies this Actor document
	 * @defaultValue `null`
	 */
	_id?: string | null | undefined

	/**
	 * The name of this Actor
	 */
	name: string

	/**
	 * An Actor subtype which configures the system data model applied
	 */
	type: ActorDataSource["type"]

	/**
	 * An image file path which provides the artwork for this Actor
	 * @defaultValue `ActorDataConstructor.DEFAULT_ICON`
	 */
	img?: string | null | undefined

	/**
	 * The system data object which is defined by the system template.json model
	 * @defaultValue template from template.json for type or `{}`
	 */
	data?: DeepPartial<ActorDataSource["data"]> | null | undefined

	/**
	 * Default Token settings which are used for Tokens created from this Actor
	 * @defaultValue `new PrototypeTokenData({ this.data.name, this.data.img })`
	 */
	token?: PrototypeTokenDataConstructorData | null | undefined

	/**
	 * A Collection of Item embedded Documents
	 * @defaultValue `new EmbeddedCollection(ItemData, [], BaseItem.implementation)`
	 */
	items?: ConstructorParameters<ConfiguredDocumentClass<typeof documents.BaseItem>>[0][] | null | undefined

	/**
	 * A collection of ActiveEffect embedded Documents
	 * @defaultValue `new EmbeddedCollection(ActiveEffectData, [], BaseActiveEffect.implementation)`
	 */
	effects?: ConstructorParameters<ConfiguredDocumentClass<typeof documents.BaseActiveEffect>>[0][] | null | undefined

	/**
	 * The _id of a Folder which contains this Actor
	 * @defaultValue `null`
	 */
	folder?: InstanceType<ConfiguredDocumentClass<typeof documents.BaseFolder>> | string | null | undefined

	/**
	 * The numeric sort value which orders this Actor relative to its siblings
	 * @defaultValue `0`
	 */
	sort?: number | null | undefined

	/**
	 * An object which configures user permissions to this Actor
	 * @defaultValue `{ default: CONST.DOCUMENT_OWNERSHIP_LEVELS.NONE }`
	 */
	permission?: Partial<Record<string, foundry.CONST.DOCUMENT_OWNERSHIP_LEVELS>> | null | undefined

	/**
	 * An object of optional key/value flags
	 * @defaultValue `{}`
	 */
	flags?: ConfiguredFlags<"Actor"> | null | undefined
}

type ActorDataBaseSource = PropertiesToSource<ActorDataBaseProperties>
type ActorDataProperties = ActorDataBaseProperties & ConfiguredData<"Actor">
type ActorDataSource = ActorDataBaseSource & ConfiguredSource<"Actor">